Move MultiCroppa component into utils
